Move auth data access functions
